refactor: extract positioning stats accumulator
